django-backend/requirements.txt
asgiref==3.6.0
Django==4.1.4
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
install==1.3.5
Markdown==3.4.1
PyJWT==2.6.0
pytz==2022.7
sqlparse==0.4.3
